Split widget: insight card
